add func_attr_printf in :
log.c
message.c
strings.c

fixed some printf warnings in:

src/nvim/undo.c
src/nvim/eval.c
src/nvim/eval/encode.c
src/nvim/eval/typval.c
src/nvim/ex_getln.c
src/nvim/fileio.c
src/nvim/lua/executor.c
src/nvim/main.c
src/nvim/regexp_nfa.c
src/nvim/shada.c
src/nvim/spellfile.c
src/nvim/tui/terminfo.c
src/nvim/garray.h

+++ src/nvim/eval/encode.c
@@ -609,14 +609,14 @@ static inline int convert_to_json_string(garray_T *const gap,
           if (ch > 0x7F && shift == 1) {
             emsgf(_("E474: String \"%.*s\" contains byte that does not start "
                     "any UTF-8 character"),
-                  utf_len - (i - shift), utf_buf + i - shift);
+                  (int)(utf_len - (i - shift)), utf_buf + i - shift);
             xfree(tofree);
             return FAIL;
           } else if ((SURROGATE_HI_START <= ch && ch <= SURROGATE_HI_END)
                      || (SURROGATE_LO_START <= ch && ch <= SURROGATE_LO_END)) {
             emsgf(_("E474: UTF-8 string contains code point which belongs "
                     "to a surrogate pair: %.*s"),
-                  utf_len - (i - shift), utf_buf + i - shift);
+                  (int)(utf_len - (i - shift)), utf_buf + i - shift);
             xfree(tofree);
             return FAIL;
           } else if (ENCODE_RAW(ch)) {
